planting seeds

+++ src/config.rs
@@ -1,6 +1,6 @@
 pub static DEBUG: bool = false;
-pub static MAP_WIDTH: u16 = 500;
-pub static MAP_HEIGHT: u16 = 500;
+pub static MAP_WIDTH: usize = 500;
+pub static MAP_HEIGHT: usize = 500;
 
 pub static SEED_PRICE: u32 = 3;
 pub static ROCK_PRICE: u32 = 1;
